Implemented basic code for fpga test

+++ floyd-warshall/src/Algorithms/BlockedFloydWarshall.cs
@@ -10,7 +10,7 @@ public static class BlockedFloydWarshall
   private readonly struct ParallelMessage
   {
     public readonly CountdownEvent Event;
-    public readonly long[] Matrix;
+    public readonly int[] Matrix;
     public readonly int OffsetX;
     public readonly int OffsetY;
     public readonly int OffsetZ;
@@ -19,7 +19,7 @@ public static class BlockedFloydWarshall
 
     public ParallelMessage(
       CountdownEvent Event,
-      long[] Matrix,
+      int[] Matrix,
       int OffsetX,
       int OffsetY,
       int OffsetZ,
@@ -48,7 +48,7 @@ public static class BlockedFloydWarshall
   public static void ParallelVectorOptimisations(Matrix.Blocks blocks)
     => ParallelVectorOptimisations(blocks.Data, blocks.Count, blocks.Size);
 
-  private static void Baseline(long[] matrix, int block_count, int block_size)
+  private static void Baseline(int[] matrix, int block_count, int block_size)
   {
     var lineral_block_size = block_size * block_size;
     var lineral_block_row_size = block_count * lineral_block_size;
@@ -57,7 +57,7 @@ public static class BlockedFloydWarshall
     {
       var offset_mm = m * lineral_block_row_size + m * lineral_block_size;
 
-      var mm = new Span<long>(matrix, offset_mm, lineral_block_size);
+      var mm = new Span<int>(matrix, offset_mm, lineral_block_size);
 
       Procedure(mm, mm, mm, block_size);
 
@@ -68,8 +68,8 @@ public static class BlockedFloydWarshall
           var offset_im = i * lineral_block_row_size + m * lineral_block_size;
           var offset_mi = m * lineral_block_row_size + i * lineral_block_size;
       
-          var im = new Span<long>(matrix, offset_im, lineral_block_size);
-          var mi = new Span<long>(matrix, offset_mi, lineral_block_size);
+          var im = new Span<int>(matrix, offset_im, lineral_block_size);
+          var mi = new Span<int>(matrix, offset_mi, lineral_block_size);
 
           Procedure(im, im, mm, block_size);
           Procedure(mi, mm, mi, block_size);
@@ -81,7 +81,7 @@ public static class BlockedFloydWarshall
         {
           var offset_im = i * lineral_block_row_size + m * lineral_block_size;
 
-          var im = new Span<long>(matrix, offset_im, lineral_block_size);
+          var im = new Span<int>(matrix, offset_im, lineral_block_size);
 
           for (var j = 0; j < block_count; ++j) 
           {
@@ -90,8 +90,8 @@ public static class BlockedFloydWarshall
               var offset_ij = i * lineral_block_row_size + j * lineral_block_size;
               var offset_mj = m * lineral_block_row_size + j * lineral_block_size;
       
-              var ij = new Span<long>(matrix, offset_ij, lineral_block_size);
-              var mj = new Span<long>(matrix, offset_mj, lineral_block_size);
+              var ij = new Span<int>(matrix, offset_ij, lineral_block_size);
+              var mj = new Span<int>(matrix, offset_mj, lineral_block_size);
 
               Procedure(ij, im, mj, block_size);
             }
@@ -101,7 +101,7 @@ public static class BlockedFloydWarshall
     }
 
     static void Procedure(
-      Span<long> ij, Span<long> ik, Span<long> kj, int block_size)
+      Span<int> ij, Span<int> ik, Span<int> kj, int block_size)
     {
       for (var k = 0; k < block_size; ++k)
       {
@@ -120,7 +120,7 @@ public static class BlockedFloydWarshall
     }
   }
 
-  private static void ParallelOptimisation(long[] matrix, int block_count, int block_size)
+  private static void ParallelOptimisation(int[] matrix, int block_count, int block_size)
   {
     var iteration_sync = new CountdownEvent(0);
 
@@ -131,7 +131,7 @@ public static class BlockedFloydWarshall
     {
       var offset_mm = m * lineral_block_row_size + m * lineral_block_size;
 
-      var mm = new Span<long>(matrix, offset_mm, lineral_block_size);
+      var mm = new Span<int>(matrix, offset_mm, lineral_block_size);
 
       Procedure(mm, mm, mm, block_size);
 
@@ -182,16 +182,16 @@ public static class BlockedFloydWarshall
     static void ParallelProcedure(
       ParallelMessage message)
     {
-      var x = new Span<long>(message.Matrix, message.OffsetX, message.SpanLength);
-      var y = new Span<long>(message.Matrix, message.OffsetY, message.SpanLength);
-      var z = new Span<long>(message.Matrix, message.OffsetZ, message.SpanLength);
+      var x = new Span<int>(message.Matrix, message.OffsetX, message.SpanLength);
+      var y = new Span<int>(message.Matrix, message.OffsetY, message.SpanLength);
+      var z = new Span<int>(message.Matrix, message.OffsetZ, message.SpanLength);
 
       Procedure(x, y, z, message.BlockSize);
 
       message.Event.Signal();
     }
     static void Procedure(
-      Span<long> ij, Span<long> ik, Span<long> kj, int block_size)
+      Span<int> ij, Span<int> ik, Span<int> kj, int block_size)
     {
       for (var k = 0; k < block_size; ++k)
       {
@@ -210,7 +210,7 @@ public static class BlockedFloydWarshall
     }
   }
 
-  private static void VectorOptimisation(long[] matrix, int block_count, int block_size)
+  private static void VectorOptimisation(int[] matrix, int block_count, int block_size)
   {
     var lineral_block_size = block_size * block_size;
     var lineral_block_row_size = block_count * lineral_block_size;
@@ -219,7 +219,7 @@ public static class BlockedFloydWarshall
     {
       var offset_mm = m * lineral_block_row_size + m * lineral_block_size;
 
-      var mm = new Span<long>(matrix, offset_mm, lineral_block_size);
+      var mm = new Span<int>(matrix, offset_mm, lineral_block_size);
 
       Procedure(mm, mm, mm, block_size);
 
@@ -230,8 +230,8 @@ public static class BlockedFloydWarshall
           var offset_im = i * lineral_block_row_size + m * lineral_block_size;
           var offset_mi = m * lineral_block_row_size + i * lineral_block_size;
       
-          var im = new Span<long>(matrix, offset_im, lineral_block_size);
-          var mi = new Span<long>(matrix, offset_mi, lineral_block_size);
+          var im = new Span<int>(matrix, offset_im, lineral_block_size);
+          var mi = new Span<int>(matrix, offset_mi, lineral_block_size);
 
           Procedure(im, im, mm, block_size);
           Procedure(mi, mm, mi, block_size);
@@ -243,7 +243,7 @@ public static class BlockedFloydWarshall
         {
           var offset_im = i * lineral_block_row_size + m * lineral_block_size;
 
-          var im = new Span<long>(matrix, offset_im, lineral_block_size);
+          var im = new Span<int>(matrix, offset_im, lineral_block_size);
 
           for (var j = 0; j < block_count; ++j) 
           {
@@ -252,8 +252,8 @@ public static class BlockedFloydWarshall
               var offset_ij = i * lineral_block_row_size + j * lineral_block_size;
               var offset_mj = m * lineral_block_row_size + j * lineral_block_size;
       
-              var ij = new Span<long>(matrix, offset_ij, lineral_block_size);
-              var mj = new Span<long>(matrix, offset_mj, lineral_block_size);
+              var ij = new Span<int>(matrix, offset_ij, lineral_block_size);
+              var mj = new Span<int>(matrix, offset_mj, lineral_block_size);
 
               Procedure(ij, im, mj, block_size);
             }
@@ -262,28 +262,28 @@ public static class BlockedFloydWarshall
       }
     }
     static void Procedure(
-      Span<long> ij, Span<long> ik, Span<long> kj, int block_size)
+      Span<int> ij, Span<int> ik, Span<int> kj, int block_size)
     {
       for (var k = 0; k < block_size; ++k)
       {
         for (var i = 0; i < block_size; ++i)
         {
-          var ik_vec = new Vector<long>(ik[i * block_size + k]);
+          var ik_vec = new Vector<int>(ik[i * block_size + k]);
 
           var j = 0;
-          for (; j < block_size - Vector<long>.Count; j += Vector<long>.Count)
+          for (; j < block_size - Vector<int>.Count; j += Vector<int>.Count)
           {
-            var ij_vec = new Vector<long>(ij.Slice(i * block_size + j, Vector<long>.Count));
-            var ikj_vec = new Vector<long>(kj.Slice(k * block_size + j, Vector<long>.Count)) + ik_vec;
+            var ij_vec = new Vector<int>(ij.Slice(i * block_size + j, Vector<int>.Count));
+            var ikj_vec = new Vector<int>(kj.Slice(k * block_size + j, Vector<int>.Count)) + ik_vec;
 
             var lt_vec = Vector.LessThan(ij_vec, ikj_vec);
-            if (lt_vec == new Vector<long>(-1))
+            if (lt_vec == new Vector<int>(-1))
             {
               continue;
             }
 
             var r_vec = Vector.ConditionalSelect(lt_vec, ij_vec, ikj_vec);
-            r_vec.CopyTo(ij.Slice(i * block_size + j, Vector<long>.Count));
+            r_vec.CopyTo(ij.Slice(i * block_size + j, Vector<int>.Count));
           }
 
           for (; j < block_size; ++j)
@@ -299,7 +299,7 @@ public static class BlockedFloydWarshall
     }
   }
 
-  private static void ParallelVectorOptimisations(long[] matrix, int block_count, int block_size)
+  private static void ParallelVectorOptimisations(int[] matrix, int block_count, int block_size)
   {
     var iteration_sync = new CountdownEvent(0);
 
@@ -310,7 +310,7 @@ public static class BlockedFloydWarshall
     {
       var offset_mm = m * lineral_block_row_size + m * lineral_block_size;
 
-      var mm = new Span<long>(matrix, offset_mm, lineral_block_size);
+      var mm = new Span<int>(matrix, offset_mm, lineral_block_size);
 
       Procedure(mm, mm, mm, block_size);
 
@@ -361,37 +361,37 @@ public static class BlockedFloydWarshall
     static void ParallelProcedure(
       ParallelMessage message)
     {
-      var x = new Span<long>(message.Matrix, message.OffsetX, message.SpanLength);
-      var y = new Span<long>(message.Matrix, message.OffsetY, message.SpanLength);
-      var z = new Span<long>(message.Matrix, message.OffsetZ, message.SpanLength);
+      var x = new Span<int>(message.Matrix, message.OffsetX, message.SpanLength);
+      var y = new Span<int>(message.Matrix, message.OffsetY, message.SpanLength);
+      var z = new Span<int>(message.Matrix, message.OffsetZ, message.SpanLength);
 
       Procedure(x, y, z, message.BlockSize);
 
       message.Event.Signal();
     }
     static void Procedure(
-      Span<long> ij, Span<long> ik, Span<long> kj, int block_size)
+      Span<int> ij, Span<int> ik, Span<int> kj, int block_size)
     {
       for (var k = 0; k < block_size; ++k)
       {
         for (var i = 0; i < block_size; ++i)
         {
-          var ik_vec = new Vector<long>(ik[i * block_size + k]);
+          var ik_vec = new Vector<int>(ik[i * block_size + k]);
 
           var j = 0;
-          for (; j < block_size - Vector<long>.Count; j += Vector<long>.Count)
+          for (; j < block_size - Vector<int>.Count; j += Vector<int>.Count)
           {
-            var ij_vec = new Vector<long>(ij.Slice(i * block_size + j, Vector<long>.Count));
-            var ikj_vec = new Vector<long>(kj.Slice(k * block_size + j, Vector<long>.Count)) + ik_vec;
+            var ij_vec = new Vector<int>(ij.Slice(i * block_size + j, Vector<int>.Count));
+            var ikj_vec = new Vector<int>(kj.Slice(k * block_size + j, Vector<int>.Count)) + ik_vec;
 
             var lt_vec = Vector.LessThan(ij_vec, ikj_vec);
-            if (lt_vec == new Vector<long>(-1))
+            if (lt_vec == new Vector<int>(-1))
             {
               continue;
             }
 
             var r_vec = Vector.ConditionalSelect(lt_vec, ij_vec, ikj_vec);
-            r_vec.CopyTo(ij.Slice(i * block_size + j, Vector<long>.Count));
+            r_vec.CopyTo(ij.Slice(i * block_size + j, Vector<int>.Count));
           }
 
           for (; j < block_size; ++j)
